refactor: ストア分割 — services/domain層を新設しストアをスリム化
Phase 2: Store decomposition

New files:
- services/weather.ts: fetchWeatherData, getWeatherCondition
- services/pin-api.ts: createPinFromStorageUrl, uploadPinWithAudio, fetchNearbyPins
- services/analysis.ts: callBackendAnalysis, uploadAudioToStorage, generateClassificationResults
- domain/geo.ts: calculateDistanceKm, calculateDistanceMeters, generateTimeTag
- domain/pin-converter.ts: convertApiResponseToPin, convertDbPinToSoundPin

Modified:
- useSoundPinStore.ts: 994→280行（-72%）、ロジックをservices/domainに分離
- useInferenceStore.ts: 524→236行（-55%）、分析ロジックをservices/analysisに分離
